added only index and confirm prompt

+++ gk/remove.nu
@@ -8,29 +8,32 @@ export def main [
   user: string@get_local_user  # The name of the owner of the repository
   repo: string@get_local_repo  # The repository you want to remove
   --noconfirm(-n)              # do not ask for confirmation
+  --only_index(-i)             # Remove only the index and leave the phyisical git alone
 ] {
 
   let index = (get index)
-
-  try {
-    $index
-    | if not ($user in $in) {
-      return "you dont have any repositories from this user"
-    } else {$in}
-    | get $user
-    | if not ($repo in $in.name) {
-      return "you dont have that repository installed"
-    } else {$in}
-    | where name == $repo
-    | get path.0
-    | rm -r $in
-  } catch {|e|
-    match $e {
-      {msg: "File(s) not found"} => {
-        print "The folder seems to be already removed, removing the index"
-      },
-      $x => {
-        return $"There was an error:\n($x)"
+  if (not $only_index) {
+    try {
+      if (not (confirm prompt)) {return "There was no deletion"}
+      $index
+      | if not ($user in $in) {
+        return "you dont have any repositories from this user"
+      } else {$in}
+      | get $user
+      | if not ($repo in $in.name) {
+        return "you dont have that repository installed"
+      } else {$in}
+      | where name == $repo
+      | get path.0
+      | rm -r $in
+    } catch {|e|
+      match $e {
+        {msg: "File(s) not found"} => {
+          print "The folder seems to be already removed, removing the index"
+        },
+        $x => {
+          return $"There was an error:\n($x)"
+        }
       }
     }
   }
@@ -53,4 +56,15 @@ def remove_from_index [user repo] {
   }
   | to nuon
   | save -f (get index path)
+}
+
+def "confirm prompt" [path] {
+  print $"This is going to remove (ansi bold)all(ansi reset) files in [($path)]!"
+  print $"are you sure you want to proceed?"
+  ["yes" "no"] 
+  | input list
+  match $in {
+    "yes" => {true}
+    "no" => {false}
+  }
 }
